# docs/api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "科研助手 sidecar",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/deep_research/parse_upload": {
      "post": {
        "summary": "Parse Upload Ep",
        "description": "上传 PDF/DOCX/TXT → 抽取 title/摘要, 缓存全文到 project 目录.",
        "operationId": "parse_upload_ep_api_deep_research_parse_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_upload_ep_api_deep_research_parse_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deep_research/lookup_title": {
      "post": {
        "summary": "Lookup Title Ep",
        "description": "用户手输题名 → crossref/openalex 反查, 返回 {found, abstract, first_author, year, url, doi}.",
        "operationId": "lookup_title_ep_api_deep_research_lookup_title_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupTitleReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deep_research/recommend": {
      "post": {
        "summary": "Recommend Ep",
        "operationId": "recommend_ep_api_deep_research_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deep_research/stream": {
      "post": {
        "summary": "Stream Ep",
        "description": "深度调研主流: search phase 做文献检索, generate phase 做深读+合成+贡献表。",
        "operationId": "stream_ep_api_deep_research_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deep_research/followup/stream": {
      "post": {
        "summary": "Followup Ep",
        "description": "追问或改写: 将当前报告 + 追问/指令交给 LLM, 流式输出答复/改写。",
        "operationId": "followup_ep_api_deep_research_followup_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowupReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_parse_upload_ep_api_deep_research_parse_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_upload_ep_api_deep_research_parse_upload_post"
      },
...
      "FollowupReq": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "ask"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "report": {
            "type": "string",
            "title": "Report"
          },
          "references": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "References",
            "default": []
          },
          "evidence": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Evidence",
            "default": []
          },
          "english_report": {
            "type": "boolean",
            "title": "English Report",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "question",
          "report"
        ],
        "title": "FollowupReq"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LookupTitleReq": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "LookupTitleReq"
      },
...
      "RecommendReq": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "refs": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Refs"
          }
        },
        "type": "object",
        "required": [
          "question",
          "refs"
        ],
        "title": "RecommendReq"
      },
...
      "StreamReq": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "field": {
            "type": "string",
            "title": "Field",
            "default": ""
          },
          "background": {
            "type": "string",
            "title": "Background",
            "default": ""
          },
          "depth": {
            "type": "string",
            "title": "Depth",
            "default": "deep"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "filters": {
            "title": "Filters"
          },
          "phase": {
            "type": "string",
            "title": "Phase",
            "default": "search"
          },
          "references": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "References",
            "default": []
          },
          "evidence": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Evidence",
            "default": []
          },
          "deep_read_targets": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Deep Read Targets",
            "default": []
          },
          "english_report": {
            "type": "boolean",
            "title": "English Report",
            "default": false
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "StreamReq"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}